Format gb_core code

+++ gb_core/cheat.cpp
@@ -29,7 +29,7 @@ void cheat::add_cheat(const std::string &code, cpu_writecb writecb) {
     }
 }
 
-uint8_t cheat::cheat_read(uint8_t ram_bank_num, uint16_t adr, uint16_t or_value) {
+uint8_t cheat::cheat_read(uint8_t ram_bank_num, uint16_t adr, uint8_t or_value) {
     if (cheat_map.find(adr) == cheat_map.end())
     {
         return or_value;
